+++ bin/scripts/skipChosenErrors.php
@@ -2,9 +2,7 @@
 
 use Awesomite\ErrorDumper\ErrorDumper;
 
-$errorDumper = new ErrorDumper();
-$errorHandler = $errorDumper->createDevHandler(E_ALL ^ E_USER_DEPRECATED);
-$errorHandler->register();
+ErrorDumper::createDevHandler(E_ALL ^ E_USER_DEPRECATED)->register();
 
 trigger_error('Test error', E_USER_DEPRECATED);
 return 'OK';